requirements.txt
anthropic>=0.43.0
yfinance>=0.2.36
praw>=7.7.0
httpx>=0.27.0
rich>=13.7.0
jinja2>=3.1.0
python-dotenv>=1.0.0
